First Scene - somewhat done?
I am able to get to the first dialogue scene, but I'm getting errors in the p5.js sketch that I don't know how to address (because I did not write that code). It has to do with using and removing the buttons. I'm also unsure of how to get to the next dialogue screen? The screen is not changing. This might be because I started everything in the setup() function and not the draw loop, so everything's only running once. I'm going to keep messing with it and see if I can get it to work.